--- src/schema/order/index.ts
@@ -1,2 +1,0 @@
-/* eslint-disable import/prefer-default-export */
-export {default as createOrderSchema} from './createOrder.schema';
