--- src/schema/product/index.ts
@@ -1,3 +1,0 @@
-/* eslint-disable import/prefer-default-export */
-export {default as addProductSchema} from './addProduct.schema';
-export {default as addProductReview} from './addReview.schema';
